+++ lib/Flags.mli
@@ -13,9 +13,22 @@ type flags = {
   padding : padding;
   separators : bool;
   prefix_non_zero : bool;
+  suffix : bool;
   zero_printing : zero_printing;
 }
 (** [flags] are passed to pretty printing functions to customize the output. *)
 
 val default : flags
 (** A default set of flags. *)
+
+val make_flags : ?zero_padding:bool -> ?left_padding:bool -> ?separators:bool -> ?prefix:bool -> ?suffix:bool -> ?zero_special:bool -> unit -> flags
+(** A function for creating flags. The [left_padding] prameter is ignored if
+    [zero_padding] is set to [true]. The default values are as follows.
+{ul {- [zero_padding=true]}
+    {- [left_padding=false]}
+    {- [separators=true]}
+    {- [prefix=true]}
+    {- [suffix=true]}
+    {- [zero_special=false]}
+}
+*)